v2

+++ static/js/typewrite.js
@@ -1,10 +1,11 @@
-var app = document.getElementById('heading');
+if ($('#heading').length) {
+  var app = document.getElementById('heading')
 
-var typewriter = new Typewriter(app, {
+  var typewriter = new Typewriter(app, {
     loop: true
-});
+  })
 
-typewriter.typeString('static site generators.')
+  typewriter.typeString('static site generators.')
     .pauseFor(2500)
     .deleteAll()
     .typeString('continuous integration.')
@@ -15,30 +16,30 @@ typewriter.typeString('static site generators.')
     .deleteAll()
     .typeString('automated testing.')
     .pauseFor(2500)
-    .start();
-
+    .start()
+}
 
 $(document).ready(function () {
-    $("#myCarousel").on("slide.bs.carousel", function (e) {
-        var $e = $(e.relatedTarget);
-        var idx = $e.index();
-        var itemsPerSlide = 3;
-        var totalItems = $(".carousel-item").length;
+  $('#myCarousel').on('slide.bs.carousel', function (e) {
+    var $e = $(e.relatedTarget)
+    var idx = $e.index()
+    var itemsPerSlide = 3
+    var totalItems = $('.carousel-item').length
 
-        if (idx >= totalItems - (itemsPerSlide - 1)) {
-            var it = itemsPerSlide - (totalItems - idx);
-            for (var i = 0; i < it; i++) {
+    if (idx >= totalItems - (itemsPerSlide - 1)) {
+      var it = itemsPerSlide - (totalItems - idx)
+      for (var i = 0; i < it; i++) {
                 // append slides to end
-                if (e.direction == "left") {
-                    $(".carousel-item")
+        if (e.direction == 'left') {
+          $('.carousel-item')
                         .eq(i)
-                        .appendTo(".carousel-inner");
-                } else {
-                    $(".carousel-item")
+                        .appendTo('.carousel-inner')
+        } else {
+          $('.carousel-item')
                         .eq(0)
-                        .appendTo($(this).find(".carousel-inner"));
-                }
-            }
+                        .appendTo($(this).find('.carousel-inner'))
         }
-    });
-});
+      }
+    }
+  })
+})
